@SARUXSARU - add currency UP/DOWN boolean pointer field (if up -> true)

+++ currencyhub/src/main/java/com/truejacobg/currencyhub/currency/dto/CurrencyDTO.java
@@ -13,7 +13,7 @@ public class CurrencyDTO {
     private String currencyCode;
     private String currencyName;
     private String currencyTable;
-    private Double mid;
+    private Float mid;
 
 
 }
